manage nan value
previsously, we were inefficiently searching the first non nan value in the future.

if a nan is detected and remove_nan is true, the new code replaces it by the previous value.

if the starting value is a nan, the feed can start by (many) 0 because it is initialized as an array of zeros. But those starting zeros are actually nan
Using np.isfinite, the code searches in the future for the first non nan value in the feed and replaces the starting zeros by this value

adding first_non_nan_value, first_non_nan_index and starting_by_nan in the feed signature